[cosmos] minor changes following the review
