--- core/src/wallet/cosmos/CosmosLikeMsgFees.hpp
@@ -1,62 +1,0 @@
-/*
- *
- * CosmosLikeMsgFees
- *
- * Created by Remi Barjon on 30/03/2020.
- *
- * The MIT License (MIT)
- *
- * Copyright (c) 2020 Ledger
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- */
-
-#ifndef LEDGER_CORE_COSMOSLIKEMSGFEES_H
-#define LEDGER_CORE_COSMOSLIKEMSGFEES_H
-
-#include <api/CosmosLikeAmount.hpp>
-
-#include <string>
-
-namespace ledger {
-namespace core {
-namespace internal {
-
-struct CosmosLikeMsgFees final
-{
-    std::string payerAddress;
-    api::CosmosLikeAmount fees;
-
-    CosmosLikeMsgFees(std::string payerAddress,
-                      api::CosmosLikeAmount fees)
-        : payerAddress(std::move(payerAddress))
-        , fees(std::move(fees))
-    {}
-
-    CosmosLikeMsgFees() = default;
-};
-
-
-} // namespace internal
-} // namespace core
-} // namespace ledger
-
-#endif // LEDGER_CORE_COSMOSLIKEMSGFEES_H
-
